+++ api-server/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file documents changes to the Deconf API Server.
 
+## next
+
+- **fix** clear legacy content cache when new content is uploaded
+
 ## 0.1.2
 
 - **new** Stage and diff assets during upsert-schedule & link them to people